pyproject.toml
[tool.poetry]
name = "library"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

fastapi = "^0.95.0"
uvicorn = "^0.21.1"
psycopg2-binary = "^2.9.5"
pyjwt = "^2.6.0"
bcrypt = "^4.0.1"
passlib = "^1.7.4"
python-dotenv = "^1.0.0"
email-validator = "^1.3.1"
python-multipart = "^0.0.6"
celery = "^5.2.7"
redis = "^4.5.4"

# analytics
python-docx = "^0.8.11"
matplotlib = "^3.7.1"
numpy = "^1.24.2"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.300"
pytest = "^7.2.2"
pre-commit = "^3.2.1"
ipython = "^8.11.0"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
